# Bring custom object store commands in parity with custom raw string commands.