--- a/libs/server/Custom/CustomCommandRegistration.cs
+++ b/libs/server/Custom/CustomCommandRegistration.cs
@@ -37,6 +37,10 @@ namespace Garnet.server.Custom
         public CommandType CommandType { get; set; }
 
         public long ExpirationTicks { get; set; }
+
+        public string ObjectCommandName { get; set; }
+
+        public CustomObjectFunctions ObjectCommand { get; set; }
     }
 
     /// <summary>
@@ -60,7 +64,7 @@ namespace Garnet.server.Custom
 
             if (instance is CustomObjectFactory cof && args is RegisterCmdArgs cofa)
             {
-                return new RegisterCustomObjectFactoryProvider(cof, cofa);
+                return new RegisterCustomObjectCommandProvider(cof, cofa);
             }
 
             if (instance is CustomTransactionProcedure ctp && args is RegisterTxnArgs ctpa)
@@ -198,6 +202,24 @@ namespace Garnet.server.Custom
         public override void Register(CustomCommandManager customCommandManager)
         {
             customCommandManager.Register(this.RegisterArgs.Name, this.RegisterArgs.NumParams, this.RegisterArgs.CommandType, this.Instance, this.RegisterArgs.CommandInfo);
+        }
+    }
+
+    /// <summary>
+    /// CustomObjectFactory registration provider
+    /// </summary>
+    internal sealed class RegisterCustomObjectCommandProvider : RegisterCustomCmdProvider<CustomObjectFunctions>
+    {
+        private CustomObjectFactory factory;
+
+        public RegisterCustomObjectCommandProvider(CustomObjectFactory instance, RegisterCmdArgs args) : base(args.ObjectCommand, args)
+        {
+            this.factory = instance;
+        }
+
+        public override void Register(CustomCommandManager customCommandManager)
+        {
+            customCommandManager.Register(RegisterArgs.Name, RegisterArgs.NumParams, RegisterArgs.CommandType, factory, RegisterArgs.ObjectCommand, RegisterArgs.CommandInfo);
         }
     }