This change brings in support for CustomObjectFunctions which provides high level of control when authoring custom object commands similar to that of custom raw string commands that already exists.

**Benchmark (.net 8)**

main branch

| Method       | Mean     | Error   | StdDev  | Allocated |
|------------- |---------:|--------:|--------:|----------:|
| ZAddRem      | 190.6 us | 1.58 us | 1.48 us |     23 KB |
| MyDictSetGet | 270.6 us | 1.04 us | 0.81 us |     30 KB |

This PR

| Method       | Mean     | Error   | StdDev  | Allocated |
|------------- |---------:|--------:|--------:|----------:|
| ZAddRem      | 192.7 us | 0.86 us | 0.76 us |     23 KB |
| MyDictSetGet | 275.3 us | 1.63 us | 1.53 us |     30 KB |